Added FastNLO writer, merging of SmallNLO grids (current strategy is just adding grids as a placeholder) and reconstruction of previously stripped scale dependence grids

+++ smallnlo/src/lib.rs
@@ -3,5 +3,5 @@ pub mod table;
 mod util;
 mod xsec;
 
-pub type Float = f32;
+pub type Float = f64;
 pub use xsec::FastNLOEvalutator;
